refactor: 채팅 컴포넌트 분리
채팅 컴포넌트 최상단에서 채팅 목록 렌더링 부분을 새로운 컴포넌트로 분리
